# src/docs/openapi.yaml
openapi: 3.0.3
info:
  title: PlaceMux API
  version: 1.0.0-mock
  description: >
    Mock implementation of the PlaceMux placement platform API.
    Response shapes and status codes are contractual and will not change
    when real persistence replaces the mock data source.
    See ARCHITECTURE.md for full conventions.
servers:
  - url: http://localhost:3000/api/v1
    description: Local development
...
paths:
...
  /students/{id}:
    get:
      tags: [Students]
      summary: Get one student
      parameters:
        - { in: path, name: id, required: true, schema: { type: string }, example: stu_001 }
      responses:
        '200': { description: The student }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      tags: [Students]
      summary: Update a student
      parameters:
        - { in: path, name: id, required: true, schema: { type: string }, example: stu_001 }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/StudentUpdate' }
      responses:
        '200': { description: Updated }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
...
  /companies/{id}:
    get:
      tags: [Companies]
      summary: Get one company
      parameters:
        - { in: path, name: id, required: true, schema: { type: string }, example: cmp_001 }
      responses:
        '200': { description: The company }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      tags: [Companies]
      summary: Update a company
      parameters:
        - { in: path, name: id, required: true, schema: { type: string }, example: cmp_001 }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CompanyUpdate' }
      responses:
        '200': { description: Updated }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
...
components:
...
  schemas:
...
    StudentUpdate:
      type: object
      properties:
        name:           { type: string }
        branch:         { type: string }
        graduationYear: { type: integer }
        cgpa:           { type: number }
        skills:         { type: array, items: { type: string } }
        status:         { type: string, enum: [seeking, placed, inactive] }
...
    CompanyUpdate:
      type: object
      properties:
        name:          { type: string }
        industry:      { type: string }
        location:      { type: string }
        employeeCount: { type: integer, minimum: 0 }
        verified:      { type: boolean }
...
    Error:
      type: object
      properties:
        success: { type: boolean, example: false }
        error:
          type: object
          properties:
            code:    { type: string, example: RESOURCE_NOT_FOUND }
            message: { type: string }
            details: { type: array, items: { type: object } }
  responses:
...
    NotFound:
      description: Resource or route does not exist
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    ValidationFailed:
      description: Body failed field-level validation
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }